Refactor EditContractPage/EditContractForm
EditContractForm previously extended the Form class which doesn't have
markup associated with it. This required that it knew the markup of the
class using it.
This goes against the wicket best practices. Thus EditContractForm now
extends Panel and contains the form.
Along with this the class has been cleaned up in general to make it
easier to understand/maintain.